Bug Fix for mlx_plot function
After 2.0 update, the resolution of the drawing canvas for the mlx_plot was increased. However, increasing the figure size substantially BEFORE the drawing process had a substantial negative impact on the performance due to how the plot_decision_regions function works within MLxtend library.
The function works by running inference on every single pixel within the provided empty figure. increasing the empty figure size, resulted in increasing the inference process exponentially.
The fix simply removes the increase of the figure size, keeping it as default. The figure size is increased AFTER the drawing is done, saving computation cost and time in the inference process.